Working on basic task generator.

+++ lib/tap/generator/destroy.rb
@@ -5,21 +5,25 @@ module Tap
         actions.reverse_each {|action| yield(action) }
       end
       
-      def directory(path)
+      def directory(path, options={})
+        path = File.expand_path(path, target_dir)
+        
         if File.exists?(path)
           log_relative :rm, path
-          added_files << File.expand_path(path)
-          rmdir(path)
+          file_task.added_files << File.expand_path(path)
+          file_task.rmdir(path) unless pretend
         else
           log_relative :missing, path
         end
       end
     
-      def file(path)
+      def file(path, options={})
+        path = File.expand_path(path, target_dir)
+        
         if File.exists?(path)
           log_relative :rm, path
-          added_files << File.expand_path(path)
-          rm(path)
+          file_task.added_files << File.expand_path(path)
+          file_task.rm(path) unless pretend
         else
           log_relative :missing, path
         end
